requirements.txt
alembic==1.7.6
cchardet==2.1.7
click==8.0.4
et-xmlfile==1.1.0
Flask==2.0.3
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.2
importlib-resources==5.4.0
itsdangerous==2.1.0
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.1.0
numpy==1.22.2
openpyxl==3.0.9
pandas==1.4.1
Pillow==9.0.1
pyodbc==4.0.32
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.32
Werkzeug==2.0.3
zipp==3.7.0
